cpu: make PcCountPair use 64 bit unsigned int for count
In PcCountPair param, change the type for "count" from 32 bit int to
64 bit unsigned int.

+++ src/cpu/probes/pc_count_tracker_manager.hh
@@ -54,7 +54,7 @@ class PcCountTrackerManager : public SimObject {
     /** a counter that stores all the target PC addresses and the number
      * of times the target PC has been executed
      */
-    std::unordered_map<Addr, int> counter;
+    std::unordered_map<Addr, uint64_t> counter;
 
     /** a set that stores all the PC Count pairs that should raise an
      * exit event at
